Small guideline for the "Lernmodus" implemented within the PopUp. Commands alert.setTitle and alert.setContentText adjusted within the method showCsvUploadDialog.
Confirmation if right or false changed to label that shows up below the buttons in red if false and in green if correct. Alos implemented a new Button "Weiter" to continue with next question. Logic implemented within the onSubmitClick method.

Timer implemented. Label created in lernmodus_layout. Methods setupTimer, updateTimer, startTimer and stopTimer implemented in LernmodusController and called methods at the correct time to simulate the timer.